Move dmarc functions out of client.py
